renamed src/db-model to src/models
