+++ src/models/settlementWindow/settlementWindowStateChange.js
